+++ src/LiveComponent/tests/Unit/Attribute/AsLiveComponentTest.php
@@ -59,5 +59,6 @@ final class AsLiveComponentTest extends TestCase
 
         $this->assertTrue(AsLiveComponent::isActionAllowed($component, 'method1'));
         $this->assertFalse(AsLiveComponent::isActionAllowed($component, 'method2'));
+        $this->assertTrue(AsLiveComponent::isActionAllowed($component, '__invoke'));
     }
 }