+++ src/TwigComponent/src/Attribute/AsTwigComponent.php
@@ -48,7 +48,7 @@ class AsTwigComponent
         $class = new \ReflectionClass($class);
 
         if (!$attribute = $class->getAttributes(static::class, \ReflectionAttribute::IS_INSTANCEOF)[0] ?? null) {
-            throw new \InvalidArgumentException(sprintf('"%s" is not a Twig Component, did you forget to add the "%s" attribute?', $class, static::class));
+            throw new \InvalidArgumentException(sprintf('"%s" is not a Twig Component, did you forget to add the "%s" attribute?', $class->getName(), static::class));
         }
 
         return $attribute->newInstance();